examples/discord-message-counter/requirements.txt
discord.py==1.7.1
prisma.io[aiohttp]
